requirements.txt
asgiref==3.7.2
autopep8==2.0.4
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
crispy-bootstrap4==2022.1
cryptography==41.0.4
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.57.0
django-countries==7.5.1
django-crispy-forms==2.0
django-debug-toolbar==3.5.0
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
pep8==1.7.1
Pillow==10.0.1
pycodestyle==2.11.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
stripe==6.7.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.6
